Implemented ChessServer\Socket\SendToTrait
