+++ src/Socket/SendToTrait.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace ChessServer\Socket;
+
+use ChessServer\Game\PlayMode;
+
+trait SendToTrait
+{
+    public function sendToOne(int $resourceId, array $res): void
+    {
+        if (isset($this->clients[$resourceId])) {
+            $this->clients[$resourceId]->send(json_encode($res));
+
+            $this->log->info('Sent message', [
+                'id' => $resourceId,
+                'cmd' => array_keys($res),
+            ]);
+        }
+    }
+
+    public function sendToMany(array $resourceIds, array $res): void
+    {
+        foreach ($resourceIds as $resourceId) {
+            $this->clients[$resourceId]->send(json_encode($res));
+        }
+
+        $this->log->info('Sent message', [
+            'ids' => $resourceIds,
+            'cmd' => array_keys($res),
+        ]);
+    }
+
+    public function sendToAll(): void
+    {
+        $res = [
+            'broadcast' => [
+                'onlineGames' => $this->gameModeStorage
+                    ->decodeByPlayMode(PlayMode::STATUS_PENDING, PlayMode::SUBMODE_ONLINE),
+            ],
+        ];
+
+        foreach ($this->clients as $client) {
+            $client->send(json_encode($res));
+        }
+    }
+}
